// prj/donation/hapi/donation-hapi-04-api/src/types/donation-types.ts
export interface User {
  firstName: string;
  lastName: string;
  email: string;
  password: string;
  _id: string;
}

export interface Candidate {
  firstName: string;
  lastName: string;
  office: string;
  _id: string;
}

export interface Donation {
  amount: number;
  method: string;
  candidate: Candidate | string;
  donor: User | string;
  lat: number;
  lng: number;
}

export interface Db {
  userStore: any;
  candidateStore: any;
  donationStore: any;
}